key exchange tested

+++ LAB4/Dtos/Chat/ChatEncodingRequest.cs
@@ -4,6 +4,6 @@
     {
         public int ChatId { get; set; }
         public int UserId { get; set;}
-        public byte[] ClientsKey { get; set; }
+        public int[] ClientsKey { get; set; }
     }
 }
